# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BrowserCaptureRequest": {
        "additionalProperties": false,
        "properties": {
          "canonical_url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Canonical Url",
            "type": "string"
          },
          "caption": {
            "$ref": "#/components/schemas/CaptureCaption"
          },
          "client_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Client Version",
            "type": "string"
          },
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/CaptureMetadata"
          },
          "page_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Url"
          },
          "platform": {
            "enum": [
              "youtube",
              "ntu_kaltura"
            ],
            "title": "Platform",
            "type": "string"
          },
          "platform_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Platform Id",
            "type": "string"
          },
          "protocol_version": {
            "const": "capture.v1",
            "title": "Protocol Version",
            "type": "string"
          }
        },
        "required": [
          "protocol_version",
          "client_version",
          "platform",
          "platform_id",
          "canonical_url",
          "metadata",
          "caption",
          "content_hash"
        ],
        "title": "BrowserCaptureRequest",
        "type": "object"
      },
      "BrowserCaptureResponse": {
        "additionalProperties": false,
        "properties": {
          "capture_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Id"
          },
          "item_public_id": {
            "title": "Item Public Id",
            "type": "string"
          },
          "lifecycle": {
            "title": "Lifecycle",
            "type": "string"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "safe_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Code"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "capture_id",
          "item_public_id",
          "platform",
          "status",
          "lifecycle"
        ],
        "title": "BrowserCaptureResponse",
        "type": "object"
      },
...
      "CaptureCaption": {
        "additionalProperties": false,
        "properties": {
          "cues": {
            "items": {
              "$ref": "#/components/schemas/CaptureCue"
            },
            "maxItems": 50000,
            "title": "Cues",
            "type": "array"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "source": {
            "anyOf": [
              {
                "enum": [
                  "official_cc",
                  "auto_caption"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CaptureCaption",
        "type": "object"
      },
...
      "CaptureCue": {
        "additionalProperties": false,
        "properties": {
          "end_sec": {
            "minimum": 0.0,
            "title": "End Sec",
            "type": "number"
          },
          "start_sec": {
            "minimum": 0.0,
            "title": "Start Sec",
            "type": "number"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "start_sec",
          "end_sec",
          "text"
        ],
        "title": "CaptureCue",
        "type": "object"
      },
      "CaptureMetadata": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/CaptureChapter"
            },
            "maxItems": 1000,
            "title": "Chapters",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 50000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration_sec": {
            "anyOf": [
              {
                "maximum": 31536000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "tags": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CaptureMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PairingExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "verifier": {
            "maxLength": 128,
            "minLength": 43,
            "title": "Verifier",
            "type": "string"
          }
        },
        "required": [
          "verifier"
        ],
        "title": "PairingExchangeRequest",
        "type": "object"
      },
      "PairingExchangeResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "access_token",
          "expires_at"
        ],
        "title": "PairingExchangeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BrowserCompanionBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "__Host-kb_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Notebook Agent Web API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/browser-companion/extension/captures": {
      "post": {
        "operationId": "submit_capture_api_v1_browser_companion_extension_captures_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowserCaptureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserCaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BrowserCompanionBearer": []
          }
        ],
        "summary": "Submit Capture",
        "tags": [
          "browser-companion"
        ]
      }
    },
...
    "/api/v1/browser-companion/extension/pairings/{pairing_id}:exchange": {
      "post": {
        "operationId": "exchange_pairing_api_v1_browser_companion_extension_pairings__pairing_id__exchange_post",
        "parameters": [
          {
            "in": "path",
            "name": "pairing_id",
            "required": true,
            "schema": {
              "title": "Pairing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingExchangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Pairing",
        "tags": [
          "browser-companion"
        ]
      }
    },
...
  }
}